rydder litt i mappestruktur og setter rootappelement for modaler
