--- src/components/oppgaver/dokumentasjonEtterspurt/dokumentasjonEtterspurt.module.css
@@ -1,4 +1,0 @@
-.unorderedList {
-    list-style: none;
-    padding-inline-start: 0;
-}
