+++ src/components/vedlegg/EttersendelseView.tsx
@@ -15,9 +15,9 @@ import {isFileUploadAllowed} from "../driftsmelding/DriftsmeldingUtilities";
 import DriftsmeldingVedlegg from "../driftsmelding/DriftsmeldingVedlegg";
 import {logInfoMessage} from "../../redux/innsynsdata/loggActions";
 import {onSendVedleggClicked} from "../oppgaver/onSendVedleggClicked";
-import AddFileButton, {TextAndButtonWrapper} from "../oppgaver/AddFileButton";
+import AddFileButton, {TextAndButtonWrapper} from "./AddFileButton";
 import {v4 as uuidv4} from "uuid";
-import FileItemView from "../oppgaver/FileItemView";
+import FileItemView from "./FileItemView";
 import {setFileUploadFailedVirusCheckInBackend} from "../../redux/innsynsdata/innsynsDataActions";
 import {logButtonOrLinkClick} from "../../utils/amplitude";
 import {BodyShort, Button, Label, Loader} from "@navikt/ds-react";
