+++ src/components/vedlegg/vedlegg.module.css
@@ -1,0 +1,3 @@
+.vedleggModal {
+    min-width: 20rem;
+}
