+++ src/styles/lists.module.css
@@ -1,0 +1,4 @@
+.unorderedList {
+    list-style: none;
+    padding-inline-start: 0;
+}
